QC test V3.0.9 dev - v03
Version 3.0.9 – 2021/09/09

Feature:
- IPV6 wiht 4 warning clear

API Updates:
- pin mapping updates
- Remove unused marcos

Misc:
- Update linux tools for compatibility issue

+++ Arduino_package/hardware/libraries/WiFi/src/AmebaWiFiUdp.cpp
@@ -157,14 +157,14 @@ size_t WiFiUDP::write(const uint8_t *buffer, size_t size) {
     return size;
 }
 
-size_t WiFiUDP::writeImmediately(const uint8_t *buffer, size_t size) {
+int WiFiUDP::writeImmediately(const uint8_t *buffer, size_t size) {
     _client_sock = 0;
     serverDrv.sendtoData(_client_sock, buffer, size, peer_ip, peer_port);
 
     return size;
 }
 
-size_t WiFiUDP::writeImmediately(const uint8_t *buffer, size_t size, uint32_t peer_ip, uint16_t peer_port) {
+int WiFiUDP::writeImmediately(const uint8_t *buffer, size_t size, uint32_t peer_ip, uint16_t peer_port) {
     _client_sock = 0;
     serverDrv.sendtoData(_client_sock, buffer, size, peer_ip, peer_port);
